+++ plugins/org.ietr.preesm.core/src/org/ietr/preesm/core/scenario/ParameterValueManager.java
@@ -207,9 +207,11 @@ public class ParameterValueManager {
    */
   public void updateWith(final PiGraph graph) {
     getParameterValues().clear();
-    for (final Parameter p : graph.getAllParameters()) {
-      if (!p.isConfigurationInterface()) {
-        addParameterValue(p);
+    if (graph != null) {
+      for (final Parameter p : graph.getAllParameters()) {
+        if (!p.isConfigurationInterface()) {
+          addParameterValue(p);
+        }
       }
     }
   }