+++ plugins/org.ietr.preesm.pimm.algorithm/src/org/ietr/preesm/pimm/algorithm/checker/PiMMAlgorithmChecker.java
@@ -111,14 +111,13 @@ public class PiMMAlgorithmChecker {
       for (final Fifo f : fifoChecker.getFifoWithVoidType()) {
         final String srcActorPath = ((AbstractActor) f.getSourcePort().eContainer()).getName() + "." + f.getSourcePort().getName();
         final String tgtActorPath = ((AbstractActor) f.getTargetPort().eContainer()).getName() + "." + f.getTargetPort().getName();
-        this.warningMsgs
-            .put("Fifo between actors " + srcActorPath + " and " + tgtActorPath + " has type \"void\" (this is not supported by code generation)", f);
+        this.warningMsgs.put("Fifo between actors " + srcActorPath + " and " + tgtActorPath + " has type \"void\" (this is not supported by code generation)",
+            f);
       }
       for (final Fifo f : fifoChecker.getFifoWithZeroRates()) {
         final String srcActorPath = ((AbstractActor) f.getSourcePort().eContainer()).getName() + "." + f.getSourcePort().getName();
         final String tgtActorPath = ((AbstractActor) f.getTargetPort().eContainer()).getName() + "." + f.getTargetPort().getName();
-        this.warningMsgs.put("Fifo between actors " + srcActorPath + " and " + tgtActorPath + " has rates equal to 0 (you may have forgotten to set them)",
-            f);
+        this.warningMsgs.put("Fifo between actors " + srcActorPath + " and " + tgtActorPath + " has rates equal to 0 (you may have forgotten to set them)", f);
       }
     }
   }
@@ -137,8 +136,8 @@ public class PiMMAlgorithmChecker {
         this.errorMsgs.put("Actor " + a.getPath() + " does not have a refinement", a);
       }
       for (final Actor a : refinementChecker.getActorsWithInvalidExtensionRefinement()) {
-        this.errorMsgs
-            .put("Refinement " + a.getRefinement().getFilePath() + " of Actor " + a.getPath() + " does not have a valid extension (.h, .idl, or .pi)", a);
+        this.errorMsgs.put("Refinement " + a.getRefinement().getFilePath() + " of Actor " + a.getPath() + " does not have a valid extension (.h, .idl, or .pi)",
+            a);
       }
       for (final Actor a : refinementChecker.getActorsWithNonExistingRefinement()) {
         this.errorMsgs.put("Refinement  " + a.getRefinement().getFilePath() + " of Actor " + a.getPath() + " does not reference an existing file", a);